+++ src/Microsoft.Health.Dicom.Core/Serialization/ExportDataOptionsJsonConverter.cs
@@ -4,6 +4,7 @@
 // -------------------------------------------------------------------------------------------------
 
 using System;
+using System.Globalization;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using EnsureThat;
@@ -11,41 +12,44 @@ using Microsoft.Health.Dicom.Core.Models.Export;
 
 namespace Microsoft.Health.Dicom.Core.Serialization;
 
-internal sealed class ExportDataOptionsJsonConverter<T> : JsonConverter<ExportDataOptions<T>>
+internal sealed class ExportDataOptionsJsonConverter : JsonConverterFactory
 {
-    private readonly Func<T, Type> _getType;
+    public override bool CanConvert(Type typeToConvert)
+        => typeToConvert.IsGenericType && typeToConvert.GetGenericTypeDefinition() == typeof(ExportDataOptions<>);
 
-    public ExportDataOptionsJsonConverter(Func<T, Type> getType)
-        => _getType = EnsureArg.IsNotNull(getType, nameof(getType));
-
-    public override ExportDataOptions<T> Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+    public override JsonConverter CreateConverter(Type typeToConvert, JsonSerializerOptions options)
     {
-        ExportDataOptions intermediate = JsonSerializer.Deserialize<ExportDataOptions>(ref reader, options);
-
-        Type type = _getType(intermediate.Type);
-        return new ExportDataOptions<T>(intermediate.Type, intermediate.Settings.Deserialize(type, options));
+        Type arg = EnsureArg.IsNotNull(typeToConvert, nameof(typeToConvert)).GetGenericArguments()[0];
+        if (arg == typeof(ExportSourceType))
+            return new ExportDataOptionsJsonConverter<ExportSourceType>(MapSourceType);
+        else if (arg == typeof(ExportDestinationType))
+            return new ExportDataOptionsJsonConverter<ExportDestinationType>(MapDestinationType);
+        else
+            throw new JsonException(
+                string.Format(CultureInfo.CurrentCulture, DicomCoreResource.InvalidType, typeToConvert));
     }
 
-    public override void Write(Utf8JsonWriter writer, ExportDataOptions<T> value, JsonSerializerOptions options)
-    {
-        writer.WriteStartObject();
+    private static Type MapSourceType(ExportSourceType type)
+        => type switch
+        {
+            ExportSourceType.Identifiers => typeof(IdentifierExportOptions),
+            _ => throw new JsonException(
+                string.Format(
+                    CultureInfo.CurrentCulture,
+                    DicomCoreResource.UnexpectedValue,
+                    nameof(ExportDataOptions<ExportSourceType>.Type),
+                    nameof(ExportSourceType.Identifiers))),
+        };
 
-        writer.WritePropertyName(GetPropertyName(nameof(ExportDataOptions<T>.Type), options.PropertyNamingPolicy));
-        JsonSerializer.Serialize(writer, value.Type, options);
-
-        writer.WritePropertyName(GetPropertyName(nameof(ExportDataOptions<T>.Settings), options.PropertyNamingPolicy));
-        JsonSerializer.Serialize(writer, value.Settings, _getType(value.Type), options);
-
-        writer.WriteEndObject();
-    }
-
-    private static string GetPropertyName(string name, JsonNamingPolicy policy)
-        => policy != null ? policy.ConvertName(name) : name;
-
-    private sealed class ExportDataOptions
-    {
-        public T Type { get; set; }
-
-        public JsonElement Settings { get; set; }
-    }
+    private static Type MapDestinationType(ExportDestinationType type)
+        => type switch
+        {
+            ExportDestinationType.AzureBlob => typeof(AzureBlobExportOptions),
+            _ => throw new JsonException(
+                string.Format(
+                    CultureInfo.CurrentCulture,
+                    DicomCoreResource.UnexpectedValue,
+                    nameof(ExportDataOptions<ExportDestinationType>.Type),
+                    nameof(ExportDestinationType.AzureBlob))),
+        };
 }
